First working test. The NetworkTester class creates a neural network that learns how to recognize an XOR of two inputs (which are always 1 or 0).
Interesting observations:
Using a linear activation tends to result in numbers moving toward infinity; eventually, all neuron outputs become NaN. This doesn't make sense to me since the XOR inputs directly determine the output.
Adding the sum of the two XOR inputs as a feature seems like it ought to be helpful, but it tends to cause either a false positive in the 1,1 case or a negative in one of the 1,0 cases.
Every once in a while, the network fails to learn, and all predictions converge.